backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
websockets==12.0
python-dotenv==1.0.1
pydantic-settings==2.2.1
supabase==2.4.2
httpx==0.27.0
groq==0.9.0
deepgram-sdk==3.2.7